For fun, I wanted to explore a different approach for node finding and parent/ancestors computations on Nodes, and start a discussion around it.

The idea is to implement a general-purpose method `node.getParentMap` that returns an `OrderedMap<key, parentNode>` of all their descendants. Then a lot of other methods reuse this map and become trivial or more efficient.

I was running the benchmarks, but then gave up on them once I realized that running the same benchmark twice could randomly show ±80% changes with no reason.

What I like about this approach is that we can memoize this single method, and skip memoization for a lot of methods that become trivial after relying on `getParentMap`.

`getParentMap` could return a `Map` instead of `OrderedMap`, which has better performance, but then `getKeysAsArray` cannot benefit from it.