@zhujinxuan 2018-04-09
Great idea, this is what I want for a long time! 
Could you post some benchmark test result?