@Soreine 2018-04-09
As I said, I could, but benchmarks look unreliable to me for now.
Let's see what comes out of https://github.com/ianstormtaylor/slate/pull/1765