@zhujinxuan 2018-04-11
Hi, I am currently working on the same file:
https://github.com/ianstormtaylor/slate/pull/1771

I do not find logical conflicts between our two PRs(thought there would be some conflicts in the memorize part).  But to ensure there is really no conflicts, could you have a look to ensure that our changed functions would not influence each other?