@zhujinxuan 2018-04-27
I think we shall use a chain-list for getAncestors.  Calling getAncestors includes a lot of insert (and construction new List by List().unshift()).  I am not sure what happens to the real performance of that, but, in benchmark tests, they are slow (theoretically it is O(logN), but see https://github.com/facebook/immutable-js/issues/584#issuecomment-131883133 and https://hypirion.com/musings/understanding-persistent-vector-pt-1).