@benjycui 2018-04-28
I don't think this PR will improve the performance.

Before this PR, we don't need to traverse the whole tree to get parent of a node(only edge case need it). But after this PR, we must traverse the whole tree, even in this use case:

```jsx
<value>
  <document>
    <paragraph>
      ...
    </paragraph>
    ...
  </document>
<value>

document.getParent(paragraph)
```