@benjycui 2018-04-28
This PR may improve performance of some use cases, but it will slow down more use cases, too.

So, I don't think that it is a good solution.