@Soreine 2018-04-29
> Before this PR, we don't need to traverse the whole tree to get parent of a node

If you're lucky, yes. But on average you need to traverse half the tree. I would even say that, statistically, when editing, you're most likely to edit nodes at the end of the document, which would push the average closer to traversing the whole tree most of the time.

If we always compute the parent map, we are being 2 times slower on average, because we always traverse the whole tree. But all future method calls can then have access to that memoized parent map. So the next calls to `getParent` for example, even for different keys, will only have to do `parentMap.get(key)` which is `O(log32 N)` instead of traversing the tree again which is `O(N)`.

![image](https://user-images.githubusercontent.com/5644953/39404360-056dfb60-4b92-11e8-84aa-2f42c4288740.png)

So this PR should improve performance overall, by adding some cost upfront. It should:

- Reduce performance for single calls on a particular value. This is not a big deal, since doing a single call only takes a few milliseconds anyway.
- Increase performance for several calls on the same value (or values that share a lot of identical nodes).