@Soreine 2018-04-29
I have run @zhujinxuan new benchmarks...

```diff
  benchmarks
    changes
      delete-backward
        user : 462.55 → 370.12 ops/sec
        hr : 746.15 → 598.01 ops/sec
      delete-forward
        user : 141.86 → 119.07 ops/sec
        hr : 197.00 → 168.89 ops/sec
      insert-text-by-key-multiple
+        user : 68.47 → 122.27 ops/sec (44% faster)
+        hr : 80.80 → 156.86 ops/sec (48% faster)
      insert-text-by-key
        user : 555.40 → 694.99 ops/sec
        hr : 775.49 → 1015.33 ops/sec
      insert-text
-        user : 2331.45 → 1538.00 ops/sec (52% slower)
-        hr : 3382.93 → 2035.99 ops/sec (66% slower)
      normalize
        user : 651.39 → 803.23 ops/sec
        hr : 883.62 → 1111.58 ops/sec
      split-block
        user : 327.76 → 400.85 ops/sec
        hr : 431.63 → 541.32 ops/sec
    models
      from-json-big
        user : 4.52 → 4.68 ops/sec
        hr : 4.71 → 4.92 ops/sec
      from-json
        user : 175.53 → 209.49 ops/sec
        hr : 222.36 → 273.70 ops/sec
      get-active-marks-at-range
        user : 317.50 → 368.49 ops/sec
        hr : 449.51 → 467.64 ops/sec
      get-blocks-at-range
        user : 781.68 → 989.03 ops/sec
        hr : 1441.46 → 1700.61 ops/sec
      get-blocks
+        user : 5145.54 → 8294.02 ops/sec (38% faster)
+        hr : 8961.49 → 12672.82 ops/sec (29% faster)
      get-characters-at-range
        user : 218.58 → 268.61 ops/sec
        hr : 292.59 → 341.02 ops/sec
      get-characters
        user : 3966.61 → 4085.31 ops/sec
        hr : 5413.69 → 5334.25 ops/sec
      get-inlines-at-range
+        user : 676.48 → 1459.90 ops/sec (54% faster)
+        hr : 828.10 → 1972.26 ops/sec (58% faster)
      get-inlines
        user : 10713.96 → 10552.63 ops/sec
        hr : 17128.01 → 17359.54 ops/sec
      get-leaves
        user : 9522.91 → 9213.89 ops/sec
        hr : 17629.79 → 16738.93 ops/sec
      get-marks-at-range
        user : 209.06 → 202.95 ops/sec
        hr : 278.97 → 257.19 ops/sec
      get-marks
        user : 660.75 → 720.29 ops/sec
        hr : 854.61 → 933.75 ops/sec
      get-path
        user : 2356.74 → 2142.12 ops/sec
        hr : 3228.52 → 2843.75 ops/sec
      get-texts-at-range
-        user : 3970.62 → 2248.78 ops/sec (77% slower)
-        hr : 5928.97 → 2956.28 ops/sec (101% slower) 😟
      get-texts
        user : 10403.75 → 9926.92 ops/sec
        hr : 16988.23 → 15712.33 ops/sec
      has-node-multiple
        user : 2098.78 → 2601.75 ops/sec
        hr : 2560.69 → 3265.02 ops/sec
      has-node
-        user : 8150.63 → 3271.16 ops/sec (149% slower) 😟
-        hr : 12856.79 → 4253.41 ops/sec (202% slower) 😟
      to-json
        user : 949.26 → 924.66 ops/sec
        hr : 1299.58 → 1236.17 ops/sec
      update-node
-        user : 3375.53 → 2387.29 ops/sec (41% slower)
-        hr : 4786.83 → 3146.95 ops/sec (52% slower)
    html-serializer
      deserialize
        user : 9.29 → 8.32 ops/sec
        hr : 10.21 → 9.03 ops/sec
      serialize
        user : 306.36 → 362.49 ops/sec
        hr : 437.62 → 506.03 ops/sec
    plain-serializer
      deserialize
        user : 472.93 → 452.47 ops/sec
        hr : 511.15 → 497.17 ops/sec
      serialize
        user : 2747.93 → 2727.83 ops/sec
        hr : 3352.30 → 3277.58 ops/sec
    rendering
      normal
        user : 82.52 → 76.63 ops/sec
        hr : 101.56 → 93.15 ops/sec
```

The results are not there. Looking into this one specifically:

```diff
      has-node-multiple
        user : 2098.78 → 2601.75 ops/sec
        hr : 2560.69 → 3265.02 ops/sec
      has-node
-        user : 8150.63 → 3271.16 ops/sec (149% slower) 😟
-        hr : 12856.79 → 4253.41 ops/sec (202% slower) 😟
```

`has-node` is a lot slower because it needs to build the parent map.
But `has-node-multiple` should be way faster, after having built the parent map once. It is not that fast though... I need to do more research