@Soreine 2018-04-29
```diff
  benchmarks
    changes
      delete-backward
        user : 462.55 → 449.65 ops/sec
        hr : 746.15 → 692.48 ops/sec
      delete-forward
        user : 141.86 → 132.96 ops/sec
        hr : 197.00 → 192.18 ops/sec
      insert-text-by-key-multiple
+        user : 68.47 → 127.82 ops/sec (46% faster)
+        hr : 80.80 → 171.74 ops/sec (53% faster)
      insert-text-by-key
+        user : 555.40 → 820.74 ops/sec (32% faster)
+        hr : 775.49 → 1226.87 ops/sec (37% faster)
      insert-text
-        user : 2331.45 → 1504.49 ops/sec (55% slower)
-        hr : 3382.93 → 2150.45 ops/sec (57% slower)
      normalize
        user : 651.39 → 803.43 ops/sec
        hr : 883.62 → 1087.30 ops/sec
      split-block
+        user : 327.76 → 462.33 ops/sec (29% faster)
+        hr : 431.63 → 648.28 ops/sec (33% faster)
    models
      from-json-big
        user : 4.52 → 4.67 ops/sec
        hr : 4.71 → 4.88 ops/sec
      from-json
        user : 175.53 → 201.11 ops/sec
        hr : 222.36 → 254.81 ops/sec
      get-active-marks-at-range
+        user : 317.50 → 2749.18 ops/sec (88% faster)
+        hr : 449.51 → 4158.74 ops/sec (89% faster)
      get-blocks-at-range
+        user : 781.68 → 2840.59 ops/sec (72% faster)
+        hr : 1441.46 → 3910.59 ops/sec (63% faster)
      get-blocks
+        user : 5145.54 → 9593.77 ops/sec (46% faster)
+        hr : 8961.49 → 14101.32 ops/sec (36% faster)
      get-characters-at-range
+        user : 218.58 → 3745.46 ops/sec (94% faster)
+        hr : 292.59 → 5004.47 ops/sec (94% faster)
      get-characters
        user : 3966.61 → 4318.90 ops/sec
        hr : 5413.69 → 5263.51 ops/sec
      get-inlines-at-range
+        user : 676.48 → 2862.92 ops/sec (76% faster)
+        hr : 828.10 → 4448.66 ops/sec (81% faster)
      get-inlines
        user : 10713.96 → 11699.45 ops/sec
        hr : 17128.01 → 19913.59 ops/sec
      get-leaves
        user : 9522.91 → 11672.89 ops/sec
        hr : 17629.79 → 17964.79 ops/sec
      get-marks-at-range
+        user : 209.06 → 2728.10 ops/sec (92% faster)
+        hr : 278.97 → 4098.63 ops/sec (93% faster)
      get-marks
        user : 660.75 → 712.19 ops/sec
        hr : 854.61 → 959.57 ops/sec
      get-path
+        user : 2356.74 → 3276.68 ops/sec (28% faster)
+        hr : 3228.52 → 4835.35 ops/sec (33% faster)
      get-texts-at-range
        user : 3970.62 → 4005.19 ops/sec
        hr : 5928.97 → 5255.94 ops/sec
      get-texts
        user : 10403.75 → 11235.53 ops/sec
        hr : 16988.23 → 19409.07 ops/sec
      has-node-multiple
+        user : 2098.78 → 4724.41 ops/sec (56% faster)
+        hr : 2560.69 → 6144.56 ops/sec (58% faster)
      has-node
-        user : 8150.63 → 5589.01 ops/sec (46% slower)
-        hr : 12856.79 → 7955.02 ops/sec (62% slower)
      to-json
        user : 949.26 → 1054.04 ops/sec
        hr : 1299.58 → 1417.61 ops/sec
      update-node
        user : 3375.53 → 3778.33 ops/sec
        hr : 4786.83 → 5258.33 ops/sec
    html-serializer
      deserialize
        user : 9.29 → 9.00 ops/sec
        hr : 10.21 → 9.76 ops/sec
      serialize
        user : 306.36 → 347.80 ops/sec
        hr : 437.62 → 501.82 ops/sec
    plain-serializer
      deserialize
        user : 472.93 → 443.71 ops/sec
        hr : 511.15 → 480.37 ops/sec
      serialize
        user : 2747.93 → 3002.24 ops/sec
        hr : 3352.30 → 3722.05 ops/sec
    rendering
      normal
        user : 82.52 → 75.36 ops/sec
        hr : 101.56 → 91.60 ops/sec
```

If I just use a `Map` instead of `OrderedMap`, it seems better. Since we found that we can get rid of `getKeysAsArray`, that might be a good approach.